+++ src/main/java/cz/cvut/fel/hlusijak/simulator/stateColoringMethod/HueStateColoringMethod.java
@@ -1,17 +1,13 @@
 package cz.cvut.fel.hlusijak.simulator.stateColoringMethod;
 
+import cz.cvut.fel.hlusijak.simulator.ruleset.RuleSet;
+import javafx.scene.paint.Color;
+import javafx.scene.paint.Paint;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Random;
-
-import com.google.common.base.Preconditions;
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableList.Builder;
-
-import cz.cvut.fel.hlusijak.simulator.ruleset.RuleSet;
-import javafx.scene.paint.Paint;
-import javafx.scene.paint.Color;
 
 public class HueStateColoringMethod implements StateColoringMethod {
     private List<Paint> cached;
@@ -39,7 +35,7 @@ public class HueStateColoringMethod implements StateColoringMethod {
 
     @Override
     public List<Paint> getColors(RuleSet ruleSet) {
-        int states = ruleSet.getNumberOfStates();
+        byte states = ruleSet.getNumberOfStates();
 
         if (cached != null) {
             return cached;
@@ -49,7 +45,7 @@ public class HueStateColoringMethod implements StateColoringMethod {
 
         cached.add(Color.WHITE);
 
-        int remainingStates = states - 1;
+        byte remainingStates = (byte) (states - 1);
 
         for (int hueIndex = 0; hueIndex < remainingStates; hueIndex++) {
             Color color = Color.hsb(360.0 * ((hueIndex / (double) remainingStates + hueOffset) % 1.0), 0.8, 1.0);
